added unencrypted kernelcaches support
the problem was that Apple changed the segments layout for the kernel image. that is why the algorithm was failing only for the kernel. added support for the new __TEXT_EXEC and __DATA,__CONST segments.

please keep in mind that this has been tested on a few kernelcaches so stuff may happen.